feat: move writers to the output object
For improved discoverability the Table and JSON "writers" can now be
acccessed directly on the output that is passed to the commands.

The table renderer now renders the table using pterm instead of
lipgloss, as this library is what we have decided to use for styling /
"rich" output. This update seems to resolve #8.

BREAKING CHANGE: This replaces the writer package with the output
package and changes how to render tables and JSON.