feat: constructor for application and global flags
Instead of creating the naistrix.Application struct manually we now have
a constructor function that must be used to create a new application. It
is possible to configure the application with a set of option funcs.

The main Run() function of the application also has a set of option funcs
to customize how the application is run.

When creating the application we also now register a set of global flags
that will be available for all users of the library. The output writer
will use one of these flags to conditionally output messages based on
the verbosity level set when running the application.

Multiple examples have been added showcasing all the new behaviour,
along with tests.

BREAKING CHANGE: Manually creating naistrix.Application will no longer
work.