feat: introduce naistrix.Arguments
Instead of passing in provided arguments as a slice of strings we pass
in a struct with getters for fetching arguments by their names instead.
This makes it somewhat easier to shuffle the order of arguments during a
refactor without needing to fix numerical indexes in the command
implementation.

There is also a method to fetch all arguments as a string slice if that
is what the implementor wants to use.

If the implementor tries to access an argument that does not exist the
code causes a panic. This is to safeguard against unexpected sideeffects
in the commands.

BREAKING CHANGE: This changes the signature of a few exported func types