refactor: use the global flag struct for the config flag
Also set the default value of the config flag directly on the struct.
This will show the default value in the help output, so no need to set
it manually as a part of the usage string.

Also refactor the initializeConfig function to a receiver method on the
application itself so that we can access the output, flags and more.