refactor: inject viper instance into the application
Instead of using the global viper instance we inject one when creating
the application, and use this throughout the application.

Also make sure that the config subcommands only deal with the contents
of the confguration file, and not every setting that Viper has.

The output from the config commands has also been slightly modified to
include some colors, and also a confirmation when creating missing
directories for the configuration file upon the initial creation.

The table test in config_test.go has been replaced with a regular test
since the steps had to be ran in a specific order for the test to pass.